+++ CHANGELOG.md
@@ -1,3 +1,69 @@
+## [0.44.18](https://github.com/startupjs/startupjs/compare/v0.44.17...v0.44.18) (2022-05-16)
+
+
+### Features
+
+* **ui/Icon:** add ability to pass custom width and height for custom icon ([066023a](https://github.com/startupjs/startupjs/commit/066023a49e99d59e4468d62f1ae43b65eeef1fd3))
+
+
+
+## [0.44.17](https://github.com/startupjs/startupjs/compare/v0.44.16...v0.44.17) (2022-05-13)
+
+
+### Bug Fixes
+
+* **bundler:** fix rare usecase of npm module in web production being not inside node_modules ([38b21dd](https://github.com/startupjs/startupjs/commit/38b21ddab0a207a6111e688f95fca9a978281146))
+
+
+
+## [0.44.16](https://github.com/startupjs/startupjs/compare/v0.44.15...v0.44.16) (2022-05-13)
+
+
+### Bug Fixes
+
+* **babel-plugin-dotenv:** return condition ([#944](https://github.com/startupjs/startupjs/issues/944)) ([bc45a02](https://github.com/startupjs/startupjs/commit/bc45a025ff7f206206f0c1be790f3cf63ee0e463))
+* **react-native-gesture-handler:** fix crashing of ios app because of cycle imports ([#856](https://github.com/startupjs/startupjs/issues/856)) ([1e22da0](https://github.com/startupjs/startupjs/commit/1e22da0c4892ee2cbce54c08785f528e89b3becf))
+* **ui/Content:** use `Div` component as root instead of react-native `View` ([#945](https://github.com/startupjs/startupjs/issues/945)) ([212e880](https://github.com/startupjs/startupjs/commit/212e880625fb475736371c5a09bda2b1fb3246f8))
+
+
+
+## [0.44.15](https://github.com/startupjs/startupjs/compare/v0.44.14...v0.44.15) (2022-04-25)
+
+
+### Bug Fixes
+
+* **cli:** fix axios version. New one requires a webpack polyfill for 'url', we'll deal with it later. ([17d8c7d](https://github.com/startupjs/startupjs/commit/17d8c7d34d11476416c30aea12935e9f3529926e))
+
+
+
+## [0.44.14](https://github.com/startupjs/startupjs/compare/v0.44.13...v0.44.14) (2022-04-22)
+
+
+### Bug Fixes
+
+* **babel-plugin-rn-stylename-to-style:** Fix search for the react component function. This fixes using css part attribute within loops and nested arrow functions ([d860627](https://github.com/startupjs/startupjs/commit/d860627a494145d9d6f212debc8441b2f6b0ee0d))
+
+
+
+## [0.44.13](https://github.com/startupjs/startupjs/compare/v0.44.12...v0.44.13) (2022-04-21)
+
+
+### Bug Fixes
+
+* **ui/NumberInput:** fix input responsiveness ([1f55bfe](https://github.com/startupjs/startupjs/commit/1f55bfec0ade49b4f9364b3a3b8f36318d199c84))
+
+
+
+## [0.44.12](https://github.com/startupjs/startupjs/compare/v0.44.11...v0.44.12) (2022-04-21)
+
+
+### Bug Fixes
+
+* **auth:** add better validatation for email field when user reset a password ([5e449b7](https://github.com/startupjs/startupjs/commit/5e449b7cbcbd60b5bd5ce887116412cbbb9edfb4))
+* **ui/NumberInput:** fix responsiveness of input wrapper ([660b714](https://github.com/startupjs/startupjs/commit/660b7149fd9373e51e844205acfad4cc77727514))
+
+
+
 ## [0.44.11](https://github.com/startupjs/startupjs/compare/v0.44.10...v0.44.11) (2022-04-19)